Split up CollectibleDetailsModal from CollectibleDetails to allow for redux state
