+++ src/containers/collectibles/components/CollectibleDetails.tsx
@@ -1,178 +1,32 @@
 import React, { useCallback, useEffect, useRef, useState } from 'react'
 
-import {
-  Button,
-  ButtonSize,
-  ButtonType,
-  IconImage,
-  IconLink,
-  LogoEth,
-  LogoSol,
-  Modal
-} from '@audius/stems'
 import cn from 'classnames'
-import { useDispatch, useSelector } from 'react-redux'
+import { useDispatch } from 'react-redux'
 
-import { ReactComponent as IconVolume } from 'assets/img/iconVolume.svg'
-import { ReactComponent as IconMute } from 'assets/img/iconVolume0.svg'
 import { ReactComponent as IconPlay } from 'assets/img/pbIconPlay.svg'
 import { useModalState } from 'common/hooks/useModalState'
-import { Chain } from 'common/models/Chain'
 import { Collectible, CollectibleMediaType } from 'common/models/Collectible'
-import { getAccountUser } from 'common/store/account/selectors'
 import { setCollectible } from 'common/store/ui/collectible-details/slice'
-import { formatDateWithTimezoneOffset } from 'common/utils/timeUtil'
-import Drawer from 'components/drawer/Drawer'
 import LoadingSpinner from 'components/loading-spinner/LoadingSpinner'
 import PerspectiveCard from 'components/perspective-card/PerspectiveCard'
 import PreloadImage from 'components/preload-image/PreloadImage'
-import Tooltip from 'components/tooltip/Tooltip'
-import { MountPlacement } from 'components/types'
-import { collectibleMessages } from 'containers/collectibles/components/CollectiblesPage'
-import styles from 'containers/collectibles/components/CollectiblesPage.module.css'
-import { MIN_COLLECTIBLES_TIER } from 'containers/profile-page/ProfilePageProvider'
-import { useFlag } from 'containers/remote-config/hooks'
-import { useSelectTierInfo } from 'containers/user-badges/hooks'
-import { badgeTiers } from 'containers/user-badges/utils'
-import { useScript } from 'hooks/useScript'
-import { FeatureFlags } from 'services/remote-config'
 import { preload } from 'utils/image'
-import { getScrollParent } from 'utils/scrollParent'
 
 import { getFrameFromGif } from '../ethCollectibleHelpers'
 
-const MODEL_VIEWER_SCRIPT_URL =
-  'https://unpkg.com/@google/model-viewer/dist/model-viewer.min.js'
-
-const CollectibleMedia: React.FC<{
-  collectible: Collectible
-  isMuted: boolean
-  toggleMute: () => void
-  isMobile: boolean
-}> = ({ collectible, isMuted, toggleMute, isMobile }) => {
-  // if it becomes possible to render more than 1 collectible detail (model or mobile drawer), then
-  // update useScript hook to handle multiple in-flight requests
-  const scriptLoaded = useScript(MODEL_VIEWER_SCRIPT_URL, true)
-
-  const { mediaType, imageUrl, videoUrl, gifUrl, threeDUrl } = collectible
-
-  const [isSvg, setIsSvg] = useState(false)
-
-  // check for svg images to give them non-empty width
-  const handleImage = useCallback(
-    (imageContainer: HTMLDivElement | null) => {
-      if (
-        mediaType === CollectibleMediaType.IMAGE &&
-        imageUrl?.endsWith('.svg') &&
-        imageContainer &&
-        getComputedStyle(imageContainer).width === '0px'
-      ) {
-        setIsSvg(true)
-      }
-    },
-    [mediaType, imageUrl, setIsSvg]
-  )
-
-  const ref3D = useRef<HTMLDivElement | null>(null)
-
-  useEffect(() => {
-    if (threeDUrl && ref3D?.current && scriptLoaded) {
-      ref3D.current.innerHTML = `<model-viewer src=${threeDUrl} auto-rotate camera-controls />`
-      const modelViewer = ref3D.current.children[0] as HTMLElement
-      modelViewer.style.minWidth = '50vw'
-      modelViewer.style.minHeight = '50vh'
-
-      if (isMobile) {
-        modelViewer.style.width = '100%'
-
-        // for 3d objects, disable parent nft drawer element scrollability if user is on 3d object
-        const scrollableAncestor = getScrollParent(modelViewer)
-        let foundDrawerAncestor = false
-        for (const item of (scrollableAncestor?.classList ?? []).values()) {
-          if (item.includes('nftDrawer')) {
-            foundDrawerAncestor = true
-            break
-          }
-        }
-        if (foundDrawerAncestor) {
-          const scrollableAncestorElement = scrollableAncestor as HTMLElement
-          const mouseEnterListener = () => {
-            scrollableAncestorElement.style.overflowY = 'hidden'
-            console.log(scrollableAncestorElement.style.overflowY)
-          }
-          const mouseLeaveListener = () => {
-            scrollableAncestorElement.style.overflowY = 'scroll'
-            console.log(scrollableAncestorElement.style.overflowY)
-          }
-          modelViewer.addEventListener('mouseenter', mouseEnterListener)
-          modelViewer.addEventListener('mouseleave', mouseLeaveListener)
-
-          return () => {
-            modelViewer.removeEventListener('mouseenter', mouseEnterListener)
-            modelViewer.removeEventListener('mouseleave', mouseLeaveListener)
-          }
-        }
-      }
-    }
-  }, [threeDUrl, ref3D, isMobile, scriptLoaded])
-
-  return mediaType === CollectibleMediaType.THREE_D ? (
-    <div className={styles.detailsMediaWrapper} ref={ref3D} />
-  ) : mediaType === CollectibleMediaType.GIF ? (
-    <div className={styles.detailsMediaWrapper}>
-      <img src={gifUrl!} alt='Collectible' />
-    </div>
-  ) : mediaType === CollectibleMediaType.VIDEO ? (
-    <div className={styles.detailsMediaWrapper} onClick={toggleMute}>
-      <video muted={isMuted} autoPlay loop playsInline src={videoUrl!}>
-        {collectibleMessages.videoNotSupported}
-      </video>
-      {isMuted ? (
-        <IconMute className={styles.volumeIcon} />
-      ) : (
-        <IconVolume className={styles.volumeIcon} />
-      )}
-    </div>
-  ) : (
-    <div
-      className={cn(styles.detailsMediaWrapper, { [styles.svg]: isSvg })}
-      ref={handleImage}
-    >
-      <img src={imageUrl!} alt='Collectible' />
-    </div>
-  )
-}
+import { collectibleMessages } from './CollectiblesPage'
+import styles from './CollectiblesPage.module.css'
 
 const CollectibleDetails: React.FC<{
   collectible: Collectible
-  isMobile: boolean
-  updateProfilePicture?: (
-    selectedFiles: any,
-    source: 'original' | 'unsplash' | 'url'
-  ) => void
-  onSave?: () => void
-}> = ({ collectible, isMobile, updateProfilePicture, onSave }) => {
+}> = ({ collectible }) => {
   const dispatch = useDispatch()
   const { mediaType, frameUrl, videoUrl, gifUrl, name } = collectible
 
   const [isModalOpen, setIsModalOpen] = useModalState('CollectibleDetails')
-  const [isPicConfirmModalOpen, setIsPicConfirmaModalOpen] = useState<boolean>(
-    false
-  )
-  const [isMuted, setIsMuted] = useState<boolean>(true)
   const [isLoading, setIsLoading] = useState(true)
   const [frame, setFrame] = useState(frameUrl)
   const [showSpinner, setShowSpinner] = useState(false)
-
-  const { isEnabled: isCollectibleOptionEnabledFlag } = useFlag(
-    FeatureFlags.NFT_IMAGE_PICKER_TAB
-  )
-  const accountUser = useSelector(getAccountUser)
-  const userId = accountUser?.user_id ?? 0
-  const { tierNumber } = useSelectTierInfo(userId)
-  const isCollectibleOptionEnabled =
-    isCollectibleOptionEnabledFlag &&
-    tierNumber >= badgeTiers.findIndex(t => t.tier === MIN_COLLECTIBLES_TIER)
 
   // Debounce showing the spinner for a second
   useEffect(() => {
@@ -210,15 +64,6 @@ const CollectibleDetails: React.FC<{
     setIsModalOpen(true)
   }, [collectible, dispatch, setIsModalOpen])
 
-  const handleClose = useCallback(() => {
-    dispatch(setCollectible({ collectible: null }))
-    setIsModalOpen(false)
-  }, [dispatch, setIsModalOpen])
-
-  const toggleMute = useCallback(() => {
-    setIsMuted(!isMuted)
-  }, [isMuted, setIsMuted])
-
   const handleVideo = useCallback(videoElement => {
     if (videoElement !== null) {
       const listener = () => {
@@ -235,16 +80,6 @@ const CollectibleDetails: React.FC<{
       }
     }
   }, [])
-
-  const onClickProfPicUpload = async () => {
-    const { imageUrl } = collectible
-    if (!updateProfilePicture || !onSave || imageUrl === null) return
-
-    const blob = await fetch(imageUrl).then(r => r.blob())
-    await updateProfilePicture([blob], 'url')
-    await onSave()
-    setIsPicConfirmaModalOpen(false)
-  }
 
   return (
     <div className={styles.detailsContainer}>
@@ -335,236 +170,6 @@ const CollectibleDetails: React.FC<{
         </>
         <div className={styles.nftTitle}>{collectible.name}</div>
       </PerspectiveCard>
-
-      <Modal
-        title='Collectible'
-        isOpen={isModalOpen && !isMobile}
-        onClose={handleClose}
-        showTitleHeader
-        showDismissButton
-        bodyClassName={styles.modalBody}
-        headerContainerClassName={styles.modalHeader}
-        titleClassName={styles.modalTitle}
-        allowScroll
-      >
-        <div className={styles.nftModal}>
-          <CollectibleMedia
-            collectible={collectible}
-            isMuted={isMuted}
-            toggleMute={toggleMute}
-            isMobile={isMobile}
-          />
-
-          <div className={styles.details}>
-            <div className={styles.detailsTitle}>{collectible.name}</div>
-            <div className={styles.detailsStamp}>
-              {collectible.isOwned ? (
-                <span className={styles.owned}>
-                  {collectibleMessages.owned}
-                </span>
-              ) : (
-                <span className={styles.created}>
-                  {collectibleMessages.created}
-                </span>
-              )}
-
-              {collectible.chain === Chain.Eth ? (
-                <Tooltip text='Ethereum' mount={MountPlacement.PARENT}>
-                  <LogoEth className={styles.chainIcon} />
-                </Tooltip>
-              ) : (
-                <Tooltip text='Solana' mount={MountPlacement.PARENT}>
-                  <LogoSol className={styles.chainIcon} />
-                </Tooltip>
-              )}
-            </div>
-
-            {collectible.dateCreated && (
-              <div className={styles.dateWrapper}>
-                <div className={styles.dateTitle}>Date Created:</div>
-                <div className={styles.date}>
-                  {formatDateWithTimezoneOffset(collectible.dateCreated)}
-                </div>
-              </div>
-            )}
-
-            {collectible.dateLastTransferred && (
-              <div className={styles.dateWrapper}>
-                <div className={styles.dateTitle}>Last Transferred:</div>
-                <div className={styles.date}>
-                  {formatDateWithTimezoneOffset(
-                    collectible.dateLastTransferred
-                  )}
-                </div>
-              </div>
-            )}
-
-            <div className={styles.detailsDescription}>
-              {collectible.description}
-            </div>
-
-            {collectible.externalLink && (
-              <a
-                className={styles.link}
-                href={collectible.externalLink}
-                target='_blank'
-                rel='noopener noreferrer'
-              >
-                <IconLink className={styles.linkIcon} />
-                {new URL(collectible.externalLink).hostname}
-              </a>
-            )}
-
-            {collectible.permaLink && (
-              <a
-                className={styles.link}
-                href={collectible.permaLink}
-                target='_blank'
-                rel='noopener noreferrer'
-              >
-                <IconLink className={styles.linkIcon} />
-                {collectibleMessages.linkToCollectible}
-              </a>
-            )}
-
-            {isCollectibleOptionEnabled &&
-              collectible.mediaType === CollectibleMediaType.IMAGE && (
-                <Button
-                  className={styles.profPicUploadButton}
-                  textClassName={styles.profPicUploadButtonText}
-                  iconClassName={styles.profPicUploadButtonIcon}
-                  onClick={() => {
-                    setIsModalOpen(false)
-                    setIsPicConfirmaModalOpen(true)
-                  }}
-                  text='Set As Profile Pic'
-                  type={ButtonType.COMMON_ALT}
-                  size={ButtonSize.SMALL}
-                  leftIcon={<IconImage />}
-                />
-              )}
-          </div>
-        </div>
-      </Modal>
-
-      <Modal
-        showTitleHeader
-        showDismissButton
-        headerContainerClassName={styles.modalHeader}
-        isOpen={isPicConfirmModalOpen}
-        onClose={() => setIsPicConfirmaModalOpen(false)}
-        titleClassName={styles.confirmModalTitle}
-        title={
-          <>
-            <IconImage />
-            <span>Set as Profile Pic</span>
-          </>
-        }
-      >
-        <div className={styles.confirmModalContainer}>
-          <p className={styles.confirmModalText}>
-            Are you sure you want to change your profile picture?
-          </p>
-
-          <div className={styles.confirmButtonContainer}>
-            <Button
-              className={styles.profPicConfirmButton}
-              onClick={() => setIsPicConfirmaModalOpen(false)}
-              text='Nevermind'
-              type={ButtonType.COMMON_ALT}
-            />
-            <Button
-              className={styles.profPicConfirmButton}
-              onClick={onClickProfPicUpload}
-              text='Yes'
-              type={ButtonType.PRIMARY_ALT}
-            />
-          </div>
-        </div>
-      </Modal>
-
-      <Drawer
-        isOpen={isModalOpen && isMobile}
-        onClose={handleClose}
-        isFullscreen
-      >
-        <div className={styles.nftDrawer}>
-          <CollectibleMedia
-            collectible={collectible}
-            isMuted={isMuted}
-            toggleMute={toggleMute}
-            isMobile={isMobile}
-          />
-
-          <div className={styles.details}>
-            <div className={styles.detailsTitle}>{collectible.name}</div>
-            <div className={styles.detailsStamp}>
-              {collectible.isOwned ? (
-                <span className={styles.owned}>
-                  {collectibleMessages.owned}
-                </span>
-              ) : (
-                <span className={styles.created}>
-                  {collectibleMessages.created}
-                </span>
-              )}
-
-              {collectible.chain === Chain.Eth ? (
-                <LogoEth className={styles.chainIcon} />
-              ) : (
-                <LogoSol className={styles.chainIcon} />
-              )}
-            </div>
-
-            {collectible.dateCreated && (
-              <div className={styles.dateWrapper}>
-                <div className={styles.dateTitle}>Date Created:</div>
-                <div className={styles.date}>
-                  {formatDateWithTimezoneOffset(collectible.dateCreated)}
-                </div>
-              </div>
-            )}
-
-            {collectible.dateLastTransferred && (
-              <div className={styles.dateWrapper}>
-                <div className={styles.dateTitle}>Last Transferred:</div>
-                <div className={styles.date}>
-                  {formatDateWithTimezoneOffset(
-                    collectible.dateLastTransferred
-                  )}
-                </div>
-              </div>
-            )}
-
-            <div className={styles.detailsDescription}>
-              {collectible.description}
-            </div>
-
-            {collectible.externalLink && (
-              <a
-                className={styles.link}
-                href={collectible.externalLink}
-                target='_blank'
-                rel='noopener noreferrer'
-              >
-                <IconLink className={styles.linkIcon} />
-                {new URL(collectible.externalLink).hostname}
-              </a>
-            )}
-            {collectible.permaLink && (
-              <a
-                className={styles.link}
-                href={collectible.permaLink}
-                target='_blank'
-                rel='noopener noreferrer'
-              >
-                <IconLink className={styles.linkIcon} />
-                {collectibleMessages.linkToCollectible}
-              </a>
-            )}
-          </div>
-        </div>
-      </Drawer>
     </div>
   )
 }
